requirements.txt
anyascii==0.3.2
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.2.15
django-appconf==1.0.6
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-axes==6.5.1
django-browser-reload==1.15.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==23.5
django-htmx==1.8.0
django-image-cropping==1.7
django-imagekit==5.0.0
django-modelcluster==6.3
django-permissionedforms==0.1
django-rest-swagger==2.2.0
django-taggit==4.0.0
django-tailwind==3.6.0
django-treebeard==4.7.1
django-user-agents==0.4.0
djangorestframework==3.15.2
draftjs-exporter==2.1.7
drf-spectacular==0.27.2
et-xmlfile==1.1.0
filelock==3.15.4
filetype==1.2.0
fsspec==2024.6.1
html5lib==1.1
idna==3.7
importlib_resources==6.4.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
l18n==2021.3
laces==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
openapi-codec==1.3.2
openpyxl==3.1.5
pilkit==3.0
pillow==10.4.0
pillow_heif==0.18.0
pkgutil_resolve_name==1.3.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
sympy==1.13.1
telepath==0.3.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
wagtail==5.2.6
webencodings==0.5.1
Willow==1.6.3
XlsxWriter==3.2.0
zipp==3.20.0
